requirements.txt
flask
flask_wtf
requests
mitmproxy
frida
pycryptodome
pure-python-adb
pymysql
wtforms
tqdm
adbutils
psutil